updated the grammar (more comprehensive). Extraction of chunks fixed
The issue was that basically the object (generator) since it was shared between the syntactic module and the semantic module, only one of them could use it.
To this end I made it list so that both the two modules could use it. Still unclear why after the first usage (by the syntactic module) the object was emptied.